Add Member replaces create child and VPC logic
