+++ lib/app/app.dart
@@ -7,7 +7,6 @@ import 'package:givt_app/app/injection/injection.dart';
 import 'package:givt_app/app/routes/app_router.dart';
 import 'package:givt_app/core/notification/notification.dart';
 import 'package:givt_app/features/auth/cubit/auth_cubit.dart';
-import 'package:givt_app/features/children/vpc/cubit/vpc_cubit.dart';
 import 'package:givt_app/l10n/l10n.dart';
 import 'package:givt_app/shared/bloc/infra/infra_cubit.dart';
 import 'package:givt_app/utils/utils.dart';
@@ -65,11 +64,6 @@ class _AppState extends State<App> {
           ),
           BlocProvider(
             create: (_) => InfraCubit(
-              getIt(),
-            ),
-          ),
-          BlocProvider(
-            create: (_) => VPCCubit(
               getIt(),
             ),
           ),
