+++ lib/app/routes/pages.dart
@@ -47,6 +47,7 @@ enum Pages {
   childDetails(path: 'child-details', name: 'CHILD-DETAILS'),
   editChild(path: 'edit-child', name: 'EDIT-CHILD'),
   creditCardDetail(path: 'credit-card-detail', name: 'CREDIT-CARD-DETAIL'),
+  addMember(path: 'add-member', name: 'ADD-MEMBER'),
   ;
 
   const Pages({
