+++ lib/core/network/api_service.dart
@@ -437,6 +437,26 @@ class APIService {
     }
   }
 
+  Future<bool> addMember(Map<String, dynamic> body) async {
+    final url = Uri.https(apiURL, '/givtservice/v1/profiles');
+
+    final response = await client.post(
+      url,
+      body: jsonEncode(body),
+      headers: {
+        'Content-Type': 'application/json',
+      },
+    );
+
+    if (response.statusCode >= 300) {
+      throw GivtServerFailure(
+        statusCode: response.statusCode,
+        body: jsonDecode(response.body) as Map<String, dynamic>,
+      );
+    }
+    return response.statusCode == 200;
+  }
+
   Future<bool> createChild(Map<String, dynamic> body) async {
     final url =
         Uri.https(apiURL, 'givtservice/v1/childprofile/setup-child-profile');
