+++ lib/features/children/add_member/mixins/mixins.dart
@@ -1,3 +1,2 @@
-export 'child_date_of_birth_validator.dart';
 export 'child_giving_allowance_validator.dart';
 export 'child_name_validator.dart';
