+++ lib/features/children/add_member/pages/vpc_page.dart
@@ -2,14 +2,14 @@ import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 import 'package:givt_app/core/enums/amplitude_events.dart';
-import 'package:givt_app/features/children/create_child/cubit/create_child_cubit.dart';
-import 'package:givt_app/features/children/vpc/widgets/vpc_notice_dialog.dart';
+import 'package:givt_app/features/children/add_member/cubit/add_member_cubit.dart';
+import 'package:givt_app/features/children/add_member/widgets/notice_dialog.dart';
 import 'package:givt_app/l10n/l10n.dart';
 import 'package:givt_app/utils/analytics_helper.dart';
 import 'package:givt_app/utils/app_theme.dart';
 
-class VPCSinglePage extends StatelessWidget {
-  const VPCSinglePage({super.key});
+class VPCPage extends StatelessWidget {
+  const VPCPage({super.key});
 
   @override
   Widget build(BuildContext context) {
@@ -18,7 +18,6 @@ class VPCSinglePage extends StatelessWidget {
     return Center(
       child: Column(
         mainAxisAlignment: MainAxisAlignment.spaceBetween,
-        crossAxisAlignment: CrossAxisAlignment.center,
         children: [
           RichText(
             textAlign: TextAlign.center,
@@ -69,7 +68,7 @@ class VPCSinglePage extends StatelessWidget {
                     backgroundColor: AppTheme.givtPurple,
                     showDragHandle: true,
                     useSafeArea: true,
-                    builder: (context) => const VPCNoticeDialog(),
+                    builder: (context) => const NoticeDialog(),
                   );
                 },
                 child: Row(
@@ -93,7 +92,7 @@ class VPCSinglePage extends StatelessWidget {
                 padding: const EdgeInsets.only(left: 15, right: 15),
                 child: ElevatedButton(
                   onPressed: () {
-                    context.read<CreateChildCubit>().createChild();
+                    context.read<AddMemberCubit>().createChildWithVPC();
                   },
                   child: Text(
                     'Done',
