+++ lib/features/children/add_member/widgets/notice_dialog.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:givt_app/l10n/l10n.dart';
 
-class VPCNoticeDialog extends StatelessWidget {
-  const VPCNoticeDialog({super.key});
+class NoticeDialog extends StatelessWidget {
+  const NoticeDialog({super.key});
 
   @override
   Widget build(BuildContext context) {
