--- lib/features/children/create_child/mixins/child_date_of_birth_validator.dart
@@ -1,14 +1,0 @@
-import 'package:givt_app/l10n/l10n.dart';
-
-mixin ChildDateOfBirthValidator {
-  String? validateDateOfBirth({
-    required AppLocalizations locals,
-    DateTime? dateOfBirth,
-  }) {
-    String? errorMessage;
-    if (dateOfBirth == null) {
-      errorMessage = locals.createChildDateErrorText;
-    }
-    return errorMessage;
-  }
-}
