+++ lib/features/children/edit_child/repositories/create_child_repository.dart
@@ -1,21 +1,13 @@
 import 'package:givt_app/core/network/api_service.dart';
-import 'package:givt_app/features/children/create_child/models/child.dart';
 import 'package:givt_app/features/children/edit_child/models/edit_child.dart';
 
-mixin CreateChildRepository {
-  Future<bool> createChild(Child child);
+mixin EditChildRepository {
   Future<bool> editChild(String childGUID, EditChild child);
 }
 
-class CreateChildRepositoryImpl with CreateChildRepository {
-  CreateChildRepositoryImpl(this.apiService);
+class EditChildRepositoryImpl with EditChildRepository {
+  EditChildRepositoryImpl(this.apiService);
   final APIService apiService;
-
-  @override
-  Future<bool> createChild(Child child) async {
-    final response = await apiService.createChild(child.toJson());
-    return response;
-  }
 
   @override
   Future<bool> editChild(String childGUID, EditChild child) async {
